+++ scripts/src/cloud/gcp/setup/destroy.sh
@@ -74,13 +74,13 @@ serviceAccounts:
       - roles/dns.admin
   - name: gcr-admin
     roles:
-      - projects/${project_name}/roles/gcrAdmin
+      - projects/${cloud_project}/roles/gcrAdmin
   - name: gcs-admin
     roles:
-      - projects/${project_name}/roles/gcsAdmin
+      - projects/${cloud_project}/roles/gcsAdmin
   - name: velero
     roles:
-      - projects/${project_name}/roles/velero
+      - projects/${cloud_project}/roles/velero
   - name: tekton
     roles:
       - roles/container.admin